renamed genId to generateAlphaName
