+++ src/components/base/containers/Drawer.md
@@ -46,7 +46,7 @@ style, where only one panel can be opened at a time.
 ### Advanced Example
 Use the ``noPadding`` property to remove default padding within a panel (useful for lists and tables).
 Use the ``titleAside`` property to display text or other content on the right side of the panel header.
-You can customize the key property of each Panel; if not specified the default is a zero-based array.
+You can customize the ``key`` property of each Panel; if not specified the default is a zero-based array.
 Both the Drawer and Drawer.Panel components will accept additional classNames.
 
 ```
