+++ src/components/util/generateAlphaName.js
